requirements.txt
# Astro Runtime includes the following pre-installed providers packages:
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

boto3
pandas
astronomer-cosmos
astronomer_cosmos
protobuf
apache-airflow-providers-amazon==9.16.0
apache-airflow-providers-postgres>=6.2.0
apache-airflow-task-sdk==1.1.2+astro.1
